refactor: extract shared listing filter logic and stop encoding state in test ids
