+++ packages/cms-base-layer/app/components/SwProductListingFilters.vue
@@ -4,23 +4,9 @@ import type {
   CmsElementSidebarFilter,
 } from "@shopware/composables";
 import { useCmsTranslations } from "@shopware/composables";
-import { getCategoryFilterPostFilter } from "@shopware/helpers";
 import { defu } from "defu";
-import { computed } from "vue";
-import type { ComputedRef } from "vue";
-import type { LocationQueryRaw } from "vue-router";
 
-import {
-  firstQueryValue,
-  getVisibleListingFilters,
-  toNumber,
-  useCategoryListing,
-  useProductSearchListing,
-  useRoute,
-  useRouter,
-  useSelectedListingFilters,
-} from "#imports";
-import type { Schemas, operations } from "#shopware";
+import { useListingFilters } from "#imports";
 
 const props = defineProps<{
   content?: CmsElementProductListing | CmsElementSidebarFilter;
@@ -47,206 +33,17 @@ let translations: Translations = {
 
 translations = defu(useCmsTranslations(), translations) as Translations;
 
-const route = useRoute();
-const router = useRouter();
-
 const {
-  changeCurrentSortingOrder,
-  getCurrentFilters,
-  getCurrentSortingOrder,
-  getInitialFilters,
+  currentSortingOrder,
   getSortingOrders,
-  search,
-} = isProductSearch ? useProductSearchListing() : useCategoryListing();
-
-const sidebarSelectedFilters = useSelectedListingFilters();
-
-// Only the filters this listing can actually apply - the category filter is
-// search-only, matching the post-filter and URL gates below.
-const visibleFilters = computed(() =>
-  getVisibleListingFilters(getInitialFilters.value, { isProductSearch }),
-);
-
-const showResetFiltersButton = computed<boolean>(() => {
-  if (
-    sidebarSelectedFilters.manufacturer.size !== 0 ||
-    sidebarSelectedFilters.properties.size !== 0 ||
-    (isProductSearch && sidebarSelectedFilters.categories.size !== 0) ||
-    sidebarSelectedFilters["max-price"] ||
-    sidebarSelectedFilters["min-price"] ||
-    sidebarSelectedFilters.rating ||
-    sidebarSelectedFilters["shipping-free"]
-  ) {
-    return true;
-  }
-
-  return false;
-});
-
-const searchCriteriaForRequest: ComputedRef<Schemas["ProductListingCriteria"]> =
-  computed(() => ({
-    manufacturer: [
-      ...(sidebarSelectedFilters.manufacturer as Set<string>),
-    ]?.join("|"),
-    properties: [...(sidebarSelectedFilters.properties as Set<string>)]?.join(
-      "|",
-    ),
-    // Category selection travels as a post-filter so the category
-    // aggregation itself is not reduced (faceted behavior). Search pages
-    // only: category pages never offer this filter, so a stale
-    // ?categories= param must not silently narrow them.
-    ...(isProductSearch && sidebarSelectedFilters.categories.size > 0
-      ? {
-          "post-filter": [
-            getCategoryFilterPostFilter([
-              ...(sidebarSelectedFilters.categories as Set<string>),
-            ]),
-          ],
-        }
-      : {}),
-    "min-price": sidebarSelectedFilters["min-price"] as number,
-    "max-price": sidebarSelectedFilters["max-price"] as number,
-    order: getCurrentSortingOrder.value as string,
-    "shipping-free": sidebarSelectedFilters["shipping-free"] as boolean,
-    rating: sidebarSelectedFilters.rating as number,
-    search: isProductSearch ? (firstQueryValue(route.query.search) ?? "") : "",
-    limit: toNumber(firstQueryValue(route.query.limit)) ?? 15,
-  }));
-
-const handleFilterChange = async (event: {
-  code: string;
-  value: string | number | boolean;
-}) => {
-  try {
-    const { code, value } = event;
-
-    if (
-      code === "manufacturer" ||
-      code === "properties" ||
-      code === "categories"
-    ) {
-      const filterSet = sidebarSelectedFilters[code];
-      const stringValue = String(value);
-
-      if (filterSet.has(stringValue)) {
-        filterSet.delete(stringValue);
-      } else {
-        filterSet.add(stringValue);
-      }
-    } else if (code === "min-price" || code === "max-price") {
-      sidebarSelectedFilters[code] =
-        typeof value === "number" ? value : Number(value);
-    } else if (code === "rating") {
-      sidebarSelectedFilters.rating = Number(value);
-    } else if (code === "shipping-free") {
-      sidebarSelectedFilters["shipping-free"] = Boolean(value);
-    }
-
-    await executeSearch();
-  } catch (error) {
-    console.error("Filter update failed:", error);
-  }
-};
-
-const executeSearch = async () => {
-  try {
-    // Build query directly from searchCriteriaForRequest which already has pipe-separated strings
-    const criteria = searchCriteriaForRequest.value;
-    const query: Record<string, unknown> = {};
-
-    if (criteria.manufacturer) query.manufacturer = criteria.manufacturer;
-    if (criteria.properties) query.properties = criteria.properties;
-    if (isProductSearch && sidebarSelectedFilters.categories.size > 0)
-      query.categories = [
-        ...(sidebarSelectedFilters.categories as Set<string>),
-      ].join("|");
-    if (criteria["min-price"]) query["min-price"] = criteria["min-price"];
-    if (criteria["max-price"]) query["max-price"] = criteria["max-price"];
-    if (criteria.rating) query.rating = criteria.rating;
-    if (criteria["shipping-free"])
-      query["shipping-free"] = criteria["shipping-free"];
-    if (criteria.order) query.order = criteria.order;
-    // Keep the search term in the URL so filtering on the search page does not reset the query
-    if (isProductSearch && route.query.search)
-      query.search = route.query.search;
-    // Preserve the page-size selection; rebuilding the query from scratch would
-    // otherwise drop `limit` and revert to the default on refresh/navigation.
-    if (route.query.limit) query.limit = route.query.limit;
-
-    // URL first: gating it on the request lost the filter when the call failed.
-    await router.push({
-      query: query as LocationQueryRaw,
-    });
-
-    // The listing refetches from the URL; fetching here too would double it.
-  } catch (error) {
-    console.error("Search execution failed:", error);
-  }
-};
-
-const clearFilters = () => {
-  (sidebarSelectedFilters.manufacturer as Set<string>).clear();
-  (sidebarSelectedFilters.properties as Set<string>).clear();
-  (sidebarSelectedFilters.categories as Set<string>).clear();
-  sidebarSelectedFilters["min-price"] = undefined;
-  sidebarSelectedFilters["max-price"] = undefined;
-  sidebarSelectedFilters.rating = undefined;
-  sidebarSelectedFilters["shipping-free"] = undefined;
-};
-
-const currentSortingOrder = computed({
-  get: (): string => getCurrentSortingOrder.value || "",
-  set: async (order: string): Promise<void> => {
-    try {
-      await router.push({
-        query: {
-          ...route.query,
-          order,
-        },
-      });
-
-      // Pushing is the whole job; the listing refetches from the URL.
-    } catch (error) {
-      console.error("Sorting order change failed:", error);
-    }
-  },
-});
-
-async function invokeCleanFilters() {
-  try {
-    clearFilters();
-    await executeSearch();
-  } catch (error) {
-    console.error("Clear filters failed:", error);
-  }
-}
-
-const handleSortChange = (sortKey: string) => {
-  currentSortingOrder.value = sortKey;
-};
-
-const handleRemoveFilterChip = async (chip: {
-  code: string;
-  value: string | number;
-}) => {
-  if (
-    chip.code === "properties" ||
-    chip.code === "manufacturer" ||
-    chip.code === "categories"
-  ) {
-    const filterSet = sidebarSelectedFilters[chip.code] as Set<string>;
-    filterSet.delete(String(chip.value));
-  } else if (chip.code === "price") {
-    sidebarSelectedFilters["min-price"] = undefined;
-    sidebarSelectedFilters["max-price"] = undefined;
-  } else if (chip.code === "rating") {
-    sidebarSelectedFilters.rating = undefined;
-  } else if (chip.code === "shipping-free") {
-    sidebarSelectedFilters["shipping-free"] = undefined;
-  }
-
-  await executeSearch();
-};
+  handleFilterChange,
+  handleRemoveFilterChip,
+  handleSortChange,
+  invokeCleanFilters,
+  selectedFilters: sidebarSelectedFilters,
+  showResetFiltersButton,
+  visibleFilters,
+} = useListingFilters(isProductSearch);
 </script>
 <template>
   <div>
